i2c: move OF helpers into the core
I2C of helpers used to live in of_i2c.c but experience (from SPI) shows
that it is much cleaner to have this in the core. This also removes a
circular dependency between the helpers and the core, and so we can
finally register child nodes in the core instead of doing this manually
in each driver. So, fix the drivers and documentation, too.

Change-Id: I3d92e6f3ebf6737af697eb5f50c518e4e0e5fe9c
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>